pyproject.toml
[tool.poetry]
name = "k8s-itlabs-operator"
version = "1.3.1"
description = "Operator for preparing infrastructure dependencies of services"
authors = ["itlabs.io"]

[tool.poetry.dependencies]
python = "^3.8"
kopf = "1.33.0"
kubernetes = "22.6.0"
hvac = "0.11.0"
psycopg2-binary = "2.9.3"
sentry-sdk = "1.9.8"
prometheus-client = "0.15.0"
ujson = "5.2.0"
wrapt = "1.14.1"
pytest = "7.1.1"
pytest-mock = "3.7.0"
factory-boy = "3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
